requirements.txt
pymongo
fastapi
uvicorn
motor
python-dotenv
rapidfuzz
